+++ ng-sudoku/src/app/component/app.component.ts
@@ -2,8 +2,8 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  templateUrl: '../html/app.component.html',
+  styleUrls: ['../css/app.component.css']
 })
 export class AppComponent {
   title = 'Sudoku Solver';
