+++ ng-sudoku/src/app/test/app.component.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed, async } from '@angular/core/testing';
 
-import { AppComponent } from './app.component';
+import { AppComponent } from '../component/app.component';
 
 describe('AppComponent', () => {
   beforeEach(async(() => {
